[fix] Display employees without fractions (#570)
* Display employees without fractions

Extract a localizeNumber function from localizeUnit to allow for better
control of the formatter options. In this case to render employees
without fractions.

* Avoid fractions for employees in CompanyOverview

* Avoid fractional employees in CompanyList

This component doesn't seem to be used but better update it in case it
gets brought in again.

* Extract localizeEmployeeCount function

To avoid having to pass in the same options everywhere, create a helper
function that formats employee count correctly.