[fix] Localise percentage formatting (#597)
* Localise percentage in LandingPage

Refactor RankedList to avoid it having knowledge about municipalities
and companies. This simplifies the code and avoids having hard coded
support for specific types of RankedLists.

Using an itemValueRenderer due to the way emissions are rendered, if
this changes in the future so that is rendered similarly to percentages
(value and unit in the same style) we can switch back to using
displayValue in each item as a string.

* Localise change rate on company detail page

* Localise percents in Scope3Chart

* Localise percentage in company details

* Localise percentage in MaunicipalityCard

* Localise percentage on municipality details page

* Localise percentage on scope cards

Also localise emissions with format function to ensure alignment with
other absolute emissions.

* Avoid division with zero in chart tooltip

* Don't use string interpolation for Tailwind color

Tailwind extracts class names by parsing the code and will not found
constructed strings.

https://tailwindcss.com/docs/detecting-classes-in-source-files#dynamic-class-names

* Fix some translation issues found during review

Seems these were not updated to new translations group.

Added a missing translation as well, using the already existing key.