Introduce Data Guide
This is the basic structure for a help system that allows context aware
help to be shown throughout the site. Each component can register
relevant help sections that will be displayed in the Data Guide when
they are mounted.

Having a single Data Guide that displays data info

Make the Data Guide a sidebar that can be shown or hidden.

Extract sidebar to avoid unnecessary rerendering

Avoid rerendering the context (and the entire site) when filtering in
the sidebar.

Revert changes to company overview/scope3 and App

Keep the Data Guide separate for now. We can look into integrating with
other help texts separately.

First stab at initial filter and search filter

Include a clear button in the sidebar search

Remove unused item

Rename HelpSystem to DataGuide

It's more fitting for what it is at the moment.