Fix layout of ChatHeader when data guide is open
Introduce a useContainerQuery to be able to query container sizes. This
allows us to use container queries instead of useScreenSize to decide
which code to run.

Using media queries and useScreenSize doesn't work very well when having
an optional sidebar as the screen size is always the same width but the
components might not get the same width.