Wrap tooltips in a contain-layout to work inside @container
@container is Tailwind for setting `container-type: inline-size;` which
implicitely adds layout containment. However Radix is using Floating UI
which doesn't seem to pick up on that and renders tooltips in the wrong
place. Explicetely setting it seems to fix the problem though.

Having to set `contain-layout` everytime a @container is set is a bit
cumbersome and error prone but works.

Better