Fit the dataguide better in sections and two column DataGuide on Desktop
To reduce vertical scrolling and make better use of the extra width,
this is a version of the inline data guide that that have two columns
and allows more informatio to be visible without scrolling.

The data guide was also made to fit nicely into the gray section boxes
as a section footer.

Mobile version was cleaned up to better fit the design of the site.